style(dashboard): add subheader for dashboard-editor
